+++ serde-version/examples/common/mod.rs
@@ -1,13 +1,17 @@
 use serde_version::{DeserializeVersioned, VersionMap, VersionedDeserializer};
 use std::fmt::Debug;
 
-pub fn deserialize_test<'de, T: DeserializeVersioned<'de> + PartialEq + Debug>(
+pub fn deserialize_test<
+    'de,
+    T: DeserializeVersioned<'de, VM> + PartialEq + Debug,
+    VM: VersionMap,
+>(
     input: &'de str,
     v: T,
-    version_map: &'de dyn VersionMap,
+    version_map: VM,
 ) {
     let mut ron_deserializer = ron::de::Deserializer::from_str(input).unwrap();
-    let deserializer = VersionedDeserializer::new(&mut ron_deserializer, version_map);
+    let deserializer = VersionedDeserializer::new(&mut ron_deserializer, version_map.clone());
     let de = T::deserialize_versioned(deserializer, version_map).unwrap();
 
     assert_eq!(v, de);