+++ serde-version/src/seed.rs
@@ -5,20 +5,26 @@ use serde::Deserializer;
 /// Seed wrapper:
 ///
 /// Wrap calls to deserialize with a VersionedDeserializer
-pub struct VersionedSeed<'v, S> {
+pub struct VersionedSeed<'v, S, VM> {
     seed: S,
-    version_map: &'v dyn VersionMap,
+    version_map: VM,
+    marker: std::marker::PhantomData<&'v ()>,
 }
 
-impl<'v, S> VersionedSeed<'v, S> {
-    pub fn new(seed: S, version_map: &'v dyn VersionMap) -> Self {
-        Self { seed, version_map }
+impl<'v, S, VM> VersionedSeed<'v, S, VM> {
+    pub fn new(seed: S, version_map: VM) -> Self {
+        Self {
+            seed,
+            version_map,
+            marker: std::marker::PhantomData,
+        }
     }
 }
 
-impl<'de, 'v, S> DeserializeSeed<'de> for VersionedSeed<'de, S>
+impl<'de, 'v, S, VM> DeserializeSeed<'de> for VersionedSeed<'de, S, VM>
 where
     S: DeserializeSeed<'de>,
+    VM: VersionMap,
 {
     type Value = S::Value;