+++ serde-version/src/visitor.rs
@@ -6,16 +6,18 @@ use serde::de::{DeserializeSeed, EnumAccess, MapAccess, SeqAccess, VariantAccess
 use serde::{Deserialize, Deserializer};
 
 /// Wrap a visitor to wrap seed or call specialized methods
-pub struct VersionedVisitor<'v, V> {
+pub struct VersionedVisitor<'v, V, VM> {
     visitor: V,
-    version_map: &'v dyn VersionMap,
+    version_map: VM,
+    marker: std::marker::PhantomData<&'v ()>,
 }
 
-impl<'v, V> VersionedVisitor<'v, V> {
-    pub fn new(visitor: V, version_map: &'v dyn VersionMap) -> Self {
+impl<'v, V, VM> VersionedVisitor<'v, V, VM> {
+    pub fn new(visitor: V, version_map: VM) -> Self {
         Self {
             visitor,
             version_map,
+            marker: std::marker::PhantomData,
         }
     }
 }
@@ -31,9 +33,10 @@ macro_rules! forward_visit {
         }
     }
 
-impl<'de, V> Visitor<'de> for VersionedVisitor<'de, V>
+impl<'de, V, VM> Visitor<'de> for VersionedVisitor<'de, V, VM>
 where
     V: Visitor<'de>,
+    VM: VersionMap,
 {
     type Value = V::Value;
 
@@ -103,6 +106,7 @@ where
         let visitor = VersionedVisitor {
             visitor,
             version_map: self.version_map,
+            marker: std::marker::PhantomData,
         };
         self.visitor
             .visit_seq(visitor)
@@ -117,6 +121,7 @@ where
         let visitor = VersionedVisitor {
             visitor,
             version_map: self.version_map,
+            marker: std::marker::PhantomData,
         };
         self.visitor
             .visit_map(visitor)
@@ -131,6 +136,7 @@ where
         let visitor = VersionedVisitor {
             visitor,
             version_map: self.version_map,
+            marker: std::marker::PhantomData,
         };
         self.visitor
             .visit_enum(visitor)
@@ -138,9 +144,10 @@ where
     }
 }
 
-impl<'de, V> SeqAccess<'de> for VersionedVisitor<'de, V>
+impl<'de, V, VM> SeqAccess<'de> for VersionedVisitor<'de, V, VM>
 where
     V: SeqAccess<'de>,
+    VM: VersionMap,
 {
     type Error = Error<V::Error>;
 
@@ -152,7 +159,7 @@ where
     where
         T: DeserializeSeed<'de>,
     {
-        let seed = VersionedSeed::new(seed, self.version_map);
+        let seed = VersionedSeed::new(seed, self.version_map.clone());
         self.visitor
             .next_element_seed(seed)
             .map_err(Error::DeserializeError)
@@ -163,14 +170,15 @@ where
     where
         T: Deserialize<'de>,
     {
-        <T as DeserializeVersioned<'de>>::next_element(self, self.version_map)
+        <T as DeserializeVersioned<'de, VM>>::next_element(self, self.version_map.clone())
             .map_err(|err| err.reduce())
     }
 }
 
-impl<'de, V> MapAccess<'de> for VersionedVisitor<'de, V>
+impl<'de, V, VM> MapAccess<'de> for VersionedVisitor<'de, V, VM>
 where
     V: MapAccess<'de>,
+    VM: VersionMap,
 {
     type Error = Error<V::Error>;
 
@@ -182,7 +190,7 @@ where
     where
         K: DeserializeSeed<'de>,
     {
-        let seed = VersionedSeed::new(seed, self.version_map);
+        let seed = VersionedSeed::new(seed, self.version_map.clone());
         self.visitor
             .next_key_seed(seed)
             .map_err(Error::DeserializeError)
@@ -196,7 +204,7 @@ where
     where
         S: DeserializeSeed<'de>,
     {
-        let seed = VersionedSeed::new(seed, self.version_map);
+        let seed = VersionedSeed::new(seed, self.version_map.clone());
         self.visitor
             .next_value_seed(seed)
             .map_err(Error::DeserializeError)
@@ -213,8 +221,8 @@ where
         K: DeserializeSeed<'de>,
         V2: DeserializeSeed<'de>,
     {
-        let kseed = VersionedSeed::new(kseed, self.version_map);
-        let vseed = VersionedSeed::new(vseed, self.version_map);
+        let kseed = VersionedSeed::new(kseed, self.version_map.clone());
+        let vseed = VersionedSeed::new(vseed, self.version_map.clone());
         self.visitor
             .next_entry_seed(kseed, vseed)
             .map_err(Error::DeserializeError)
@@ -225,7 +233,7 @@ where
     where
         K: Deserialize<'de>,
     {
-        <K as DeserializeVersioned<'de>>::next_key(self, self.version_map)
+        <K as DeserializeVersioned<'de, VM>>::next_key(self, self.version_map.clone())
             .map_err(|err| err.reduce())
     }
 
@@ -234,7 +242,7 @@ where
     where
         V2: Deserialize<'de>,
     {
-        <V2 as DeserializeVersioned<'de>>::next_value(self, self.version_map)
+        <V2 as DeserializeVersioned<'de, VM>>::next_value(self, self.version_map.clone())
             .map_err(|err| err.reduce())
     }
 
@@ -243,28 +251,30 @@ where
     }
 }
 
-impl<'de, V> EnumAccess<'de> for VersionedVisitor<'de, V>
+impl<'de, V, VM> EnumAccess<'de> for VersionedVisitor<'de, V, VM>
 where
     V: EnumAccess<'de>,
+    VM: VersionMap,
 {
     type Error = Error<V::Error>;
-    type Variant = VersionedVisitor<'de, V::Variant>;
+    type Variant = VersionedVisitor<'de, V::Variant, VM>;
 
     #[inline]
     #[allow(clippy::type_complexity)]
     fn variant_seed<S>(
         self,
         seed: S,
-    ) -> Result<(S::Value, VersionedVisitor<'de, V::Variant>), Self::Error>
+    ) -> Result<(S::Value, VersionedVisitor<'de, V::Variant, VM>), Self::Error>
     where
         S: DeserializeSeed<'de>,
     {
-        let seed = VersionedSeed::new(seed, self.version_map);
+        let seed = VersionedSeed::new(seed, self.version_map.clone());
         match self.visitor.variant_seed(seed) {
             Ok((value, variant)) => {
                 let variant = VersionedVisitor {
                     visitor: variant,
                     version_map: self.version_map,
+                    marker: std::marker::PhantomData,
                 };
                 Ok((value, variant))
             }
@@ -277,14 +287,16 @@ where
     where
         V2: Deserialize<'de>,
     {
-        let version_map = self.version_map;
-        <V2 as DeserializeVersioned<'de>>::variant(self, version_map).map_err(|err| err.reduce())
+        let version_map = self.version_map.clone();
+        <V2 as DeserializeVersioned<'de, VM>>::variant(self, version_map)
+            .map_err(|err| err.reduce())
     }
 }
 
-impl<'de, V> VariantAccess<'de> for VersionedVisitor<'de, V>
+impl<'de, V, VM> VariantAccess<'de> for VersionedVisitor<'de, V, VM>
 where
     V: VariantAccess<'de>,
+    VM: VersionMap,
 {
     type Error = Error<V::Error>;
 
@@ -312,6 +324,7 @@ where
         let visitor = VersionedVisitor {
             visitor,
             version_map: self.version_map,
+            marker: std::marker::PhantomData,
         };
         self.visitor
             .tuple_variant(len, visitor)
@@ -330,6 +343,7 @@ where
         let visitor = VersionedVisitor {
             visitor,
             version_map: self.version_map,
+            marker: std::marker::PhantomData,
         };
         self.visitor
             .struct_variant(fields, visitor)